FSA: Implement logic for FSA interaction with BFCache
Implements the logic for how the File System Access API behaves in
BFCache. The basic idea is that if lock is held only by pages that are
in the BFCache and an active page tries to take a lock that is in
contention with the lock, the pages that hold it are evicted and the
lock is released allowing the active page to take it.

Bug: 1382215, 1241174
Change-Id: I81f9e1b514f84836dc68780da26efe9af6d68f71