Shared Storage: Allow fetch/img header writes if permission on redirect
We correct the behavior of writing to shared storage from response
headers to match the spec with regard to how `PermissionsPolicy`
checks are handled.

Currently, if permission is revoked for any request in a redirect
chain, then no subsequent request in that chain can write to shared
storage from response headers.

This CL updates the behavior for `fetch()` and `HTMLImageElement` so
that, for redirect chains of requests that have opted-in via
`sharedStorageWritable`, each request in the chain has its
`PermissionsPolicy` checked independently of the others in the chain.

https://crrev.com/c/4935951 will do the same for `HTMLIframeElement`.

Bug:1218540,1489536
Change-Id: I3110a237fb7f960b91f5940a3309abc623e81dbc