url: Implement URL.canParse.
The WPT is adjusted slightly since it currently exercises behavior
it was presumably not intended to -- namely, the quirk we have on
Windows where URLs which begin with a single letter and a colon are
interpreted as file URLs relative to the given drive letter.

The particulars of URL parsing are adequately tested elsewhere, and
this test serves to demonstrate that URL.canParse(...) behaves the
same way that new URL(...) does.

Bug: 1425839
Change-Id: I13c425c1eaea791da2ccd2a1486112b5c2d6ad0b